pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "mcp>=0.9.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "whispey>=0.1.0",
]
